+++ projects/ngx-mat-lib/src/lib/xmat-media-query-state/index.ts
@@ -1,7 +1,8 @@
-import {NgModule} from "@angular/core";
-import {CommonModule} from "@angular/common";
+import { NgModule, ModuleWithProviders } from "@angular/core";
+import { CommonModule } from "@angular/common";
+import { MediaObserver } from "@angular/flex-layout";
 //
-import {XmatMediaQueryStateComponent} from "./xmat-media-query-state.component";
+import { XmatMediaQueryStateComponent, XmatMediaService } from "./xmat-media-query-state.component";
 
 export * from "./xmat-media-query-state.component";
 
@@ -12,9 +13,21 @@ export * from "./xmat-media-query-state.component";
     declarations: [
         XmatMediaQueryStateComponent
     ],
+    providers: [
+        XmatMediaService
+    ],
     exports: [
         XmatMediaQueryStateComponent
     ]
 })
 export class XmatMediaQueryStateModule {
+
+    static forRoot(): ModuleWithProviders {
+        return {
+            ngModule: XmatMediaQueryStateModule,
+            providers: [
+                { provide: MediaObserver, useValue: XmatMediaService, multi: true },
+            ]
+        };
+    }
 }