+++ projects/ngx-mat-lib/src/lib/xmat-models/xmat-select.model.ts
@@ -2,4 +2,3 @@ export interface XmatSelect<T = any> {
     description: string;
     value: T;
 }
-