+++ projects/ngx-mat-lib/src/lib/xmat-summary/xmat-summary.component.ts
@@ -1,13 +1,12 @@
-import {Component, Input, OnDestroy} from "@angular/core";
-import {XmatSummaryConfig} from "../xmat-models/xmat-summary.model";
-import {XmatSelect} from "../xmat-models/xmat-select.model";
-import {MediaChange, MediaObserver} from "@angular/flex-layout";
+import { Component, Input, OnDestroy } from "@angular/core";
+import { XmatSelect, XmatSummaryConfig } from "ngx-mat-lib";
+import { MediaChange, MediaObserver } from "@angular/flex-layout";
 //
-import {XmatBreakpointKeys} from "../xmat-models/xmat-breakpoints.model";
+import { XmatBreakpointKeys } from "../xmat-models/xmat-breakpoints.model";
 //
-import {chunk, merge, map, forEach} from "lodash";
-import {Subject} from "rxjs";
-import {takeUntil} from "rxjs/operators";
+import { chunk, merge, map, forEach } from "lodash";
+import { Subject } from "rxjs";
+import { takeUntil } from "rxjs/operators";
 
 @Component({
     selector: "xmat-summary",
@@ -50,15 +49,15 @@ export class XmatSummaryComponent implements OnDestroy {
 
     constructor(mediaService: MediaObserver) {
         mediaService.asObservable()
-        .pipe(takeUntil(this._subsCtrl$))
-        .subscribe((changes: MediaChange[]) => {
-            const change = changes[0];
-            const previousMqAlias = this._currentMqAlias; // Clone previous state to check for columns change
-            if (previousMqAlias !== change.mqAlias) {
-                this._currentMqAlias = change.mqAlias;
-                !!this._config && this._needsBuild(previousMqAlias, change.mqAlias) && this._buildSummary();
-            }
-        });
+            .pipe(takeUntil(this._subsCtrl$))
+            .subscribe((changes: MediaChange[]) => {
+                const change = changes[0];
+                const previousMqAlias = this._currentMqAlias; // Clone previous state to check for columns change
+                if (previousMqAlias !== change.mqAlias) {
+                    this._currentMqAlias = change.mqAlias;
+                    !!this._config && this._needsBuild(previousMqAlias, change.mqAlias) && this._buildSummary();
+                }
+            });
     }
 
     ngOnDestroy(): void {