+++ src/app/services/xmat-test-consts.service.ts
@@ -1,7 +1,7 @@
 import {Injectable} from "@angular/core";
 import {
     XmatConstantsService,
-    XmatDynamicRestVerbsRef,
+    XmatRestVerbsRef,
     XmatSelect,
     XmatGenericObject
 } from "ngx-mat-lib";